+++ src/app/highlights/page.tsx
@@ -1,0 +1,7 @@
+'use client'
+
+import { HighlightsPageView } from '@/components/highlights/HighlightsPageView'
+
+export default function HighlightsPage() {
+	return <HighlightsPageView />
+}